+++ src/frameworks/uikit/ui_view/ui_window.rs
@@ -4,11 +4,22 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 //! `UIWindow`.
+//!
+//! Useful resources:
+//! - [Technical Q&A QA1588: Automatic orientation support for iPhone and iPad apps](https://developer.apple.com/library/archive/qa/qa1588/_index.html)
+//! - [Technical Q&A QA1688: Why won't my UIViewController rotate with the device?](https://developer.apple.com/library/archive/qa/qa1688/_index.html)
 
 use super::UIViewHostObject;
 use crate::dyld::{ConstantExports, HostConstant};
+use crate::frameworks::core_graphics::cg_affine_transform::CGAffineTransform;
 use crate::frameworks::core_graphics::{CGPoint, CGRect};
 use crate::frameworks::foundation::ns_string;
+use crate::frameworks::uikit::ui_application::{
+    UIInterfaceOrientationLandscapeLeft, UIInterfaceOrientationLandscapeRight,
+};
+use crate::frameworks::uikit::ui_device::{
+    UIDeviceOrientationLandscapeLeft, UIDeviceOrientationLandscapeRight,
+};
 use crate::objc::{id, msg, msg_class, msg_super, nil, objc_classes, ClassExports};
 
 #[derive(Default)]
@@ -141,6 +152,62 @@ pub const CLASSES: ClassExports = objc_classes! {
     () = msg![env; vc viewWillAppear:false];
     () = msg_super![env; this addSubview:view];
     () = msg![env; vc viewDidAppear:false];
+
+    // Support auto-rotation. This is currently only for apps that request a
+    // non-portrait interface orientation via Info.plist, as we do not yet
+    // support changes of orientation caused by device rotation (TODO).
+    // FIXME: It's unclear when and where this auto-rotation is supposed to
+    //        happen. It must have something to do with mounting the view
+    //        controller to a window, so we do it here. QA1688 (see top of file)
+    //        mentions a breaking behaviour change in iOS 6 that makes
+    //        auto-rotation rely on rootViewController (a property only found in
+    //        iOS 6), so the current implementation is specific to iOS <= 5.
+    // FIXME: Are we supposed to notify the view somehow of the rotation?
+    // FIXME: What do we do if shouldAutorotateToInterfaceOrientation:
+    //        returns false? The status bar has already been rotated…
+    // FIXME: The device orientation stored on env.window can come from one of
+    //        three places (user/default options, setStatusBarOrientation: etc,
+    //        Info.plist UIInterfaceOrientation etc). It's not clear if these
+    //        are really equivalent and should all trigger autorotation.
+    if let Some(orientation) = match env.window.as_ref().unwrap().current_rotation() {
+        crate::window::DeviceOrientation::LandscapeLeft => Some(UIDeviceOrientationLandscapeLeft),
+        crate::window::DeviceOrientation::LandscapeRight => Some(UIDeviceOrientationLandscapeRight),
+        // Portrait is the default so we don't do anything here.
+        crate::window::DeviceOrientation::Portrait => None,
+    } {
+        // (UIInterfaceOrientation and UIDeviceOrientation are compatible enums,
+        //  here we use whichever is clearer contextually.)
+        let should = msg![env; vc shouldAutorotateToInterfaceOrientation:orientation];
+        log_dbg!("[{:?} shouldAutorotateToInterfaceOrientation:{:?}] => {:?}", vc, orientation, should);
+        if should {
+            log_dbg!("App requested autorotation; applying orientation transform to view {:?}.", view);
+            let transform = match orientation {
+                UIInterfaceOrientationLandscapeLeft => CGAffineTransform::make_rotation(-std::f32::consts::FRAC_PI_2),
+                UIInterfaceOrientationLandscapeRight => CGAffineTransform::make_rotation(std::f32::consts::FRAC_PI_2),
+                _ => unimplemented!(),
+            };
+
+            let window_frame: CGRect = msg![env; this frame];
+            log_dbg!("Window frame: {window_frame:?}");
+            let view_frame: CGRect = msg![env; view frame];
+            log_dbg!("Old view frame: {view_frame:?}");
+
+            () = msg![env; view setTransform:transform];
+
+            // Re-apply the view's old frame to compensate for the rotation
+            // effectively offseting its center position and changing the size.
+            // FIXME: I have no idea if this is how this should be solved, but
+            //        it works for DMC4 Refrain at least.
+
+            let view_frame: CGRect = msg![env; view frame];
+            log_dbg!("Old view frame after transform: {view_frame:?}");
+
+            () = msg![env; view setFrame:window_frame];
+
+            let view_frame: CGRect = msg![env; view frame];
+            log_dbg!("New view frame after re-applying old view frame: {view_frame:?}");
+        }
+    }
 }
 
 - (CGPoint)convertPoint:(CGPoint)point