+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Changes are categorised as follows:
 Compatibility:
 
 - Fixed an assumption that the backing store of a CAEAGLLayer should always be 320×480 pixels. This hack is over three years old and was missed when introducing real support for Core Animation. This should fix stretched or offset screen issues in certain games that render in landscape. (@hikari-no-yume)
+- API support improvements:
+  - Limited support for UIKit auto-rotation has been added to fix certain games that render in landscape:
+    - If an app requires a landscape orientation via the `UIInterfaceOrientation` and `UISupportedInterfaceOrientations` keys of its `Info.plist`, touchHLE will rotate the virtual device at startup. (@hikari-no-yume)
+    - If an app overrides the `shouldAutorotateToInterfaceOrientation:` method of `UIViewController`, and the virtual device is in a landscape orientation, then the root view of a view controller will be rotated when it is added to a window. (@hikari-no-yume)
 
 ## v0.2.3 (2026-01-02)